feat: add workspace analyzer to handle AI-based project profiling
