+++ extension/secureflow/.windsurf/workflows/pr.md
@@ -7,7 +7,7 @@ This workflow is dedicated for creating github pull request from command line.
 
 Follow the instruction below to create a descriptive pull request based on commits comparing to main
 
-1. Ensure you're in a branch apart from main
+1. Ensure you're in a branch apart from main / default
 2. Go through the commits in the branch comparing to main (or default branch)
 3. Go through the commit message, code changes
 4. Come up with good decent pull request message body
