+++ extension/secureflow/packages/secureflow-cli/lib/index.d.ts
@@ -10,3 +10,5 @@ export { AIModel } from './types';
 // Export prompts functionality
 export { getPromptPath, getAppProfilerPrompt } from './prompts';
 export { loadPrompt, getPromptForAppType, getApplicationProfilerPrompt, getThreatModelingPrompt } from './prompts/prompt-loader';
+
+export { WorkspaceAnalyzer, ApplicationProfile, WorkspaceAnalyzerOptions } from './workspace-analyzer';
