+++ extension/secureflow/packages/secureflow-cli/lib/workspace-analyzer/index.d.ts
@@ -1,0 +1,1 @@
+export { WorkspaceAnalyzer, ApplicationProfile, WorkspaceAnalyzerOptions } from './workspace-analyzer';
