+++ extension/secureflow/packages/secureflow-cli/lib/workspace-analyzer/index.js
@@ -1,0 +1,6 @@
+const { WorkspaceAnalyzer, ApplicationProfile } = require('./workspace-analyzer');
+
+module.exports = {
+  WorkspaceAnalyzer,
+  ApplicationProfile
+};
